+++ src/main/java/model/Guetekriterien/Guetekriterium.java
@@ -5,7 +5,7 @@ import model.Zufallszahlengenerator;
 public interface  Guetekriterium {
 
 
-    double berechneGuetekriterium(Zufallszahlengenerator generator , float[] zahlenfolge);
+    double berechneGuetekriterium(Zufallszahlengenerator generator , double[] zahlenfolge);
 
 
 }
